feat(data): gorm-gen

+++ internal/biz/user.go
@@ -9,9 +9,9 @@ import (
 
 // UserDO  领域对象（ Domain Object， DO），微服务运行时核心业务对象的载体， DO 一般包括实体或值对象。
 type UserDO struct {
-	Id        uint
-	Age       uint32
-	UserId    uint32
+	Id        int64
+	Age       int32
+	UserId    int64
 	UserName  string
 	Password  string
 	Email     string
@@ -28,11 +28,11 @@ type UserUsecase struct {
 }
 
 type UserRepo interface {
-	CreateUser(context.Context, *UserDO) (uint32, error)
+	CreateUser(context.Context, *UserDO) (int64, error)
 	UpdateUser(context.Context, *UserDO) error
-	DeleteUser(context.Context, uint32) error
-	SelectUserByID(ctx context.Context, id uint) (*UserDO, error)
-	SelectUserByUid(ctx context.Context, userId uint32) (*UserDO, error)
+	DeleteUser(context.Context, int64) error
+	SelectUserByID(ctx context.Context, id int64) (*UserDO, error)
+	SelectUserByUid(ctx context.Context, userId int64) (*UserDO, error)
 	ListUser(ctx context.Context, pageNum, pageSize uint32) ([]*UserDO, error)
 	VerifyPassword(context.Context, *UserDO) (bool, error)
 	SelectUserByEmail(context.Context, string) (*UserDO, error)
@@ -43,7 +43,7 @@ func NewUserBiz(repo UserRepo, logger log.Logger) *UserUsecase {
 	return &UserUsecase{repo: repo, log: log.NewHelper(logger)}
 }
 
-func (uc *UserUsecase) Create(ctx context.Context, user *UserDO) (uint32, error) {
+func (uc *UserUsecase) Create(ctx context.Context, user *UserDO) (int64, error) {
 	if err := uc.Validate(ctx, user); err != nil {
 		return 0, err
 	}
@@ -59,15 +59,15 @@ func (uc *UserUsecase) Update(ctx context.Context, user *UserDO) error {
 	return uc.repo.UpdateUser(ctx, user)
 }
 
-func (uc *UserUsecase) Delete(ctx context.Context, userId uint32) error {
+func (uc *UserUsecase) Delete(ctx context.Context, userId int64) error {
 	return uc.repo.DeleteUser(ctx, userId)
 }
 
-func (uc *UserUsecase) Get(ctx context.Context, id uint) (*UserDO, error) {
+func (uc *UserUsecase) Get(ctx context.Context, id int64) (*UserDO, error) {
 	return uc.repo.SelectUserByID(ctx, id)
 }
 
-func (uc *UserUsecase) GetByUID(ctx context.Context, userId uint32) (*UserDO, error) {
+func (uc *UserUsecase) GetByUID(ctx context.Context, userId int64) (*UserDO, error) {
 	return uc.repo.SelectUserByUid(ctx, userId)
 }
